move generic functions to focused util packages
